Fix undefined variable warnings
In content it seems there was either something planned but unimplemented, or some weird copy paste of a function signature. Either way default and alt variables were undefined.

Now I've set it to use the WP 'retro' option for default avatars and then the users 'display_name' and a sprintf for the alt text.